feat: add .path and .path.silent functions to detect binary
After installation if you wish to specify a path to the installed
binary, it's useful for this module to be able to tell the world
where it put the binary, so this PR adds `.path` and `.path.silent()`
functions to do that.

This way we can use the installed module without having to rely on
any external context or platform.